artefacts/checkAkv/requirements.txt
azure-identity==1.5.0
azure-keyvault-secrets==4.2.0